PR change task 1: Move django test to 1.16
