+++ django/django_test.go
@@ -2,7 +2,6 @@ package django
 
 import (
 	"bytes"
-	"embed"
 	"io/ioutil"
 	"net/http"
 	"regexp"
@@ -109,29 +108,6 @@ func Test_FileSystem(t *testing.T) {
 	}
 }
 
-//go:embed views
-var viewsAsssets embed.FS
-
-func Test_PathForwardingFileSystem(t *testing.T) {
-	engine := NewPathForwardingFileSystem(http.FS(viewsAsssets), "/views", ".django")
-	engine.Debug(true)
-	if err := engine.Load(); err != nil {
-		t.Fatalf("load: %v\n", err)
-	}
-
-	var buf bytes.Buffer
-	err := engine.Render(&buf, "descendant", map[string]interface{}{
-		"greeting": "World",
-	})
-	if err != nil {
-		t.Fatalf("render: %v", err)
-	}
-	expect := `<h1>Hello World! from ancestor</h1>`
-	result := trim(buf.String())
-	if expect != result {
-		t.Fatalf("Expected:\n%s\nResult:\n%s\n", expect, result)
-	}
-}
 
 func Test_AddFunc(t *testing.T) {
 	engine := NewFileSystem(http.Dir("./views"), ".django")
